+++ src/pyzlc/__init__.py
@@ -15,8 +15,8 @@ import concurrent.futures
 
 from .nodes.lancom_node import LanComNode
 from .nodes.nodes_info_manager import NodeInfo
-from .nodes.loop_manager import LanComLoopManager, DaemonThreadPoolExecutor, TaskReturnT
-from .sockets.service_client import zlc_request_async, zlc_request
+from .nodes.loop_manager import TaskLoopManager, DaemonThreadPoolExecutor, TaskReturnT
+from .sockets.service_client import call, async_call
 from .sockets.publisher import Publisher, Streamer
 from .utils.msg import Empty, empty, _get_zlc_version
 from .utils.log import _logger, LogLevel
@@ -66,13 +66,13 @@ def init(
     group: str = "224.0.0.1",
     group_port: int = 7720,
     log_level: Union[LogLevel, int] = LogLevel.INFO,
-    sub_group: bool = False,
+    default_group: bool = True,
 ) -> None:
     set_log_level(log_level)
     """Initialize the LanCom node singleton."""
     if LanComNode.get(group_name) is not None:
         raise ValueError("Node is already initialized.")
-    LanComNode.init(node_name, node_ip, group, group_port, group_name, sub_group)
+    LanComNode.init(node_name, node_ip, group, group_port, group_name, default_group)
     register_service_handler(
         "get_node_info",
         LanComNode.get_instance(group_name)._get_node_info_handler,
@@ -122,27 +122,7 @@ def spin(group_name: Optional[str] = None) -> None:
         LanComNode.get_instance(group_name).loop_manager.spin()
     except KeyboardInterrupt:
         _logger.debug("LanCom node interrupted by user")
-        LanComNode.get_instance(group_name).stop_node()
-
-
-def call(
-    service_name: str,
-    request: Any,
-    timeout: float = 2.0,
-    group_name: Optional[str] = None,
-) -> Any:
-    """Call a service with the specified name and request."""
-    return zlc_request(service_name, request, timeout, group_name)
-
-
-async def async_call(
-    service_name: str,
-    request: Any,
-    timeout: float = 2.0,
-    group_name: Optional[str] = None,
-) -> Any:
-    """Asynchronously call a service with the specified name and request."""
-    return await zlc_request_async(service_name, request, timeout, group_name)
+        LanComNode.stop_all_nodes()
 
 
 def register_service_handler(
@@ -162,10 +142,12 @@ def register_subscriber_handler(
     topic_name: str,
     callback: Callable,
     group_name: Optional[str] = None,
+    buffer_size: int = 1000,
+    conflate: bool = False,
 ) -> None:
     """Create a subscriber for the specified topic."""
     subscriber_manager = LanComNode.get_instance(group_name).subscriber_manager
-    subscriber_manager.add_subscriber(topic_name, callback)
+    subscriber_manager.add_subscriber(topic_name, callback, buffer_size, conflate)
 
 
 def wait_for_service(
@@ -218,7 +200,7 @@ def submit_loop_task(
     group_name: Optional[str] = None
 ) -> concurrent.futures.Future:
     """Submit a coroutine to the event loop."""
-    assert LanComLoopManager is not None, "LanComNode is not initialized."
+    assert TaskLoopManager is not None, "LanComNode is not initialized."
     return LanComNode.get_instance(group_name).loop_manager.submit_loop_task(task)
 
 
@@ -226,7 +208,7 @@ def submit_thread_pool_task(
     func: Callable[..., TaskReturnT], *args: Any
 ) -> concurrent.futures.Future:
     """Submit a synchronous function to the thread pool executor."""
-    return DaemonThreadPoolExecutor.submit_thread_pool_task(func, *args)
+    return TaskLoopManager.get_instance().submit_thread_pool_task(func, *args)
 
 
 info = _logger.info