+++ src/pyzlc/sockets/subscriber_manager.py
@@ -1,13 +1,14 @@
 import traceback
-from typing import Callable, List, Any, Optional
-import asyncio
+from typing import Callable, Dict, List, Any, Optional
 from concurrent.futures import Future
 import zmq
 import msgpack
 
+from ..utils.node_info import NodeInfo
+
 from ..nodes.zmq_socket_manager import ZMQSocketManager
 from ..utils.log import _logger
-from ..nodes.loop_manager import LanComLoopManager
+from ..nodes.loop_manager import TaskLoopManager
 from ..nodes.nodes_info_manager import NodesInfoManager
 
 class Subscriber:
@@ -17,49 +18,52 @@ class Subscriber:
         self,
         topic_name: str,
         callback: Callable[[Any], None],
-        nodes_info_manager: NodesInfoManager,
-        loop_manager: LanComLoopManager,
-        group_name: Optional[str] = None,
+        buffer_size: int = 1000,
+        conflate: bool = False,
     ):
-        self.nodes_info_manager = nodes_info_manager
-        self.loop_manager = loop_manager
         self._socket = ZMQSocketManager.get_instance().create_async_socket(zmq.SUB)
         self._socket.setsockopt_string(zmq.SUBSCRIBE, "")
+        if conflate:
+            self._socket.setsockopt(zmq.CONFLATE, 1)
+        else:
+            self._socket.setsockopt(zmq.SNDHWM, buffer_size)
         self.name = topic_name
         self.callback = callback
         self.running: bool = True
         self.connected: bool = False
-        self.published_urls: List[str] = []
-        self._listen_future: Optional[Future] = None
+        self.sub_urls: List[str] = []
+        # self._listen_future: Optional[Future] = None
         self._receive_future: Optional[Future] = None
-        self._listen_future = self.loop_manager.submit_loop_task(self.listen_loop())
+        # self._listen_future = self.loop_manager.submit_loop_task(self.listen_loop())
 
     def connect(self, url: str) -> None:
         """Connect to a publisher's socket."""
+        if url in self.sub_urls:
+            return
         self._socket.connect(url)
         self.connected = True
-        self.published_urls.append(url)
+        self.sub_urls.append(url)
         _logger.info("Subscriber %s is connected to %s", self.name, url)
-        if len(self.published_urls) == 1:
-            self._receive_future = self.loop_manager.submit_loop_task(
+        if self._receive_future is None or self._receive_future.done():
+            self._receive_future = TaskLoopManager.get_instance().submit_loop_task(
                 self.receive_loop()
             )
 
-    async def listen_loop(self) -> None:
-        """Listens for new publishers and connects to them."""
-        _logger.info("Subscriber %s is listening ...", self.name)
-        while self.running:
-            try:
-                publishers = self.nodes_info_manager.get_publisher_info(self.name)
-                for pub_info in publishers:
-                    _url = f"tcp://{pub_info['ip']}:{pub_info['port']}"
-                    if _url not in self.published_urls:
-                        self.connect(_url)
-                await asyncio.sleep(0.5)
-            except Exception as e:
-                _logger.error("Error from topic %s listener: %s", self.name, e)
-                traceback.print_exc()
-                raise e
+    # async def listen_loop(self) -> None:
+    #     """Listens for new publishers and connects to them."""
+    #     _logger.info("Subscriber %s is listening ...", self.name)
+    #     while self.running:
+    #         try:
+    #             publishers = self.nodes_info_manager.get_publisher_info(self.name)
+    #             for pub_info in publishers:
+    #                 _url = f"tcp://{pub_info['ip']}:{pub_info['port']}"
+    #                 if _url not in self.sub_urls:
+    #                     self.connect(_url)
+    #             await asyncio.sleep(0.5)
+    #         except Exception as e:
+    #             _logger.error("Error from topic %s listener: %s", self.name, e)
+    #             traceback.print_exc()
+    #             raise e
 
     async def receive_loop(self) -> None:
         """Listens for incoming messages on the subscribed topic."""
@@ -85,8 +89,6 @@ class Subscriber:
     def close(self) -> None:
         """Close the subscriber socket."""
         self.running = False
-        if self._listen_future is not None:
-            self._listen_future.cancel()
         if self._receive_future is not None:
             self._receive_future.cancel()
         self._socket.close()
@@ -96,17 +98,38 @@ class Subscriber:
 class SubscriberManager:
     """Manages multiple subscribers."""
 
-    def __init__(self, loop_manager: LanComLoopManager, nodes_info_manager: NodesInfoManager) -> None:
-        self.subscribers: List[Subscriber] = []
+    def __init__(self, loop_manager: TaskLoopManager, nodes_info_manager: NodesInfoManager) -> None:
+        # self.subscribers: List[Subscriber] = []
+        self.subscriber_dict: Dict[str, Subscriber] = {}
         self.loop_manager = loop_manager
         self.nodes_info_manager = nodes_info_manager
+        self.nodes_info_manager.register_node_update_handler("*", self.check_new_node)
 
-    def add_subscriber(self, topic_name: str, callback: Callable[[Any], None]) -> None:
+    def add_subscriber(
+        self,
+        topic_name: str,
+        callback: Callable[[Any], None],
+        buffer_size: int = 1000,
+        conflate: bool = False
+    ) -> None:
         """Add a new subscriber and start its listening and receiving loops."""
-        subscriber = Subscriber(topic_name, callback, self.nodes_info_manager, self.loop_manager)
-        self.subscribers.append(subscriber)
+        subscriber = Subscriber(topic_name, callback, buffer_size, conflate)
+        pub_infos = self.nodes_info_manager.get_publisher_info(topic_name)
+        for info in pub_infos:
+            _url = f"tcp://{info['ip']}:{info['port']}"
+            subscriber.connect(_url)
+        self.subscriber_dict[topic_name] = subscriber
 
-    def on_shutdown(self) -> None:
+    def stop(self) -> None:
         """Shutdown all subscriber sockets."""
-        for subscriber in self.subscribers:
+        self.nodes_info_manager.unregister_node_update_handler("*", self.check_new_node)
+        for subscriber in self.subscriber_dict.values():
             subscriber.close()
+
+    def check_new_node(self, node_info: NodeInfo) -> None:
+        """Check if a new node has published the subscribed topic and connect to it."""
+        for topic in node_info["topics"]:
+            if topic["name"] in self.subscriber_dict:
+                _url = f"tcp://{node_info['ip']}:{topic['port']}"
+                if _url not in self.subscriber_dict[topic["name"]].sub_urls:
+                    self.subscriber_dict[topic["name"]].connect(_url)